@@ -1,12 +1,11 @@
-﻿namespace GlaucomaWay.Models
+﻿namespace GlaucomaWay.Models;
+
+public enum Vf14Answer : short
 {
-    public enum Vf14Answer : short
-    {
-        Na = 0,
-        None = 1,
-        Alittle = 2,
-        Moderate = 3,
-        GreatDeal = 4,
-        UnableToDo = 5
-    }
+    Na = 0,
+    None = 1,
+    Alittle = 2,
+    Moderate = 3,
+    GreatDeal = 4,
+    UnableToDo = 5
 }
