@@ -2,19 +2,18 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace GlaucomaWay.Repositories.Interfaces
+namespace GlaucomaWay.Repositories.Interfaces;
+
+public interface IGenericRepository<TEntity>
+    where TEntity : class
 {
-    public interface IGenericRepository<TEntity>
-        where TEntity : class
-    {
-        Task<TEntity> CreateAsync(TEntity entity, CancellationToken cancellationToken);
+    Task<TEntity> CreateAsync(TEntity entity, CancellationToken cancellationToken);
 
-        Task DeleteAsync(object id, CancellationToken cancellationToken);
+    Task DeleteAsync(object id, CancellationToken cancellationToken);
 
-        Task<IEnumerable<TEntity>> GetAllAsync(CancellationToken cancellationToken);
+    Task<IEnumerable<TEntity>> GetAllAsync(CancellationToken cancellationToken);
 
-        Task<TEntity> GetByIdAsync(object id, CancellationToken cancellationToken);
+    Task<TEntity> GetByIdAsync(object id, CancellationToken cancellationToken);
 
-        Task UpdateAsync(TEntity obj, CancellationToken cancellationToken);
-    }
+    Task UpdateAsync(TEntity obj, CancellationToken cancellationToken);
 }
