@@ -1,11 +1,10 @@
-﻿namespace GlaucomaWay.Users
+﻿namespace GlaucomaWay.Users;
+
+public class RegisterModel
 {
-    public class RegisterModel
-    {
-        public string Username { get; set; } = string.Empty;
+    public string Username { get; set; } = string.Empty;
 
-        public string Email { get; set; } = string.Empty;
+    public string Email { get; set; } = string.Empty;
 
-        public string Password { get; set; } = string.Empty;
-    }
+    public string Password { get; set; } = string.Empty;
 }
