@@ -1,10 +1,9 @@
 ﻿using Microsoft.AspNetCore.Identity;
 
-namespace GlaucomaWay.Users
+namespace GlaucomaWay.Users;
+
+public class Role : IdentityRole
 {
-    public class Role : IdentityRole
-    {
-        public const string Admin = nameof(Admin);
-        public const string User = nameof(User);
-    }
+    public const string Admin = nameof(Admin);
+    public const string User = nameof(User);
 }
