@@ -2,13 +2,12 @@
 using GlaucomaWay.Models;
 using Microsoft.AspNetCore.Identity;
 
-namespace GlaucomaWay.Users
-{
-    public class User : IdentityUser
-    {
-        [JsonIgnore]
-        public string PatientId { get; set; }
+namespace GlaucomaWay.Users;
 
-        public PatientModel Patient { get; set; }
-    }
+public class User : IdentityUser
+{
+    [JsonIgnore]
+    public string PatientId { get; set; }
+
+    public PatientModel Patient { get; set; }
 }
